scripts/requirements.txt
beautifulsoup4==4.12.3
dashing==0.1.0
gcloud==0.18.3
google==3.0.0
google-api-core==2.19.1
google-api-python-client==2.140.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud==0.34.0
googleapis-common-protos==1.63.2
httplib2==0.22.0
jsonpointer==2.1
numpy==2.0.1
oauth2client==4.1.3
oauthlib==3.2.2
pandas==2.2.2
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.27.3
requests
rsa==4.9
soupsieve==2.5
toolz
tqdm
tzdata==2024.1
uritemplate==4.1.1
wcwidth==0.2.6
